pyproject.toml
[project]
name = "fastapi_simple_oauth2"
description = "A stateless OAuth2 middleware for FastAPI with PKCE flow support"
authors = [
    {name = "iaalm", email = "iaalmsimon@gmail.com"}
]
dynamic = [
    "version",
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "PyJWT>=2.0.0",
    "cryptography>=3.4.0",
    "python-multipart>=0.0.5"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.version]
path = "fastapi_simple_oauth2/__about__.py"

[tool.hatch.build]
exclude = [
    "tests",
]

[dependency-groups]
dev = [
    "mypy>=1.0.0",
    "httpx>=0.24.0",
    "uvicorn>=0.15.0",
]
